further work on decisions
initializations of action and service clients were moved to the
blackboard.

Paused was changed to work with a pause and unpause command card.

get_reevaluate was added to HasCustomer, so that it notices if
has_customer was set to True in the LookForCustomer action

The Idle decision pushes a sequence of actions now, instead of
just one.

CleanUp was renamed to PutBottleBack.